Refactor data access namespaces to core namespaces
The data access namespaces and interfaces have been refactored and moved to core namespaces across numerous files. This change includes the entities, DTOs, and repository interfaces and implementations. These changes lead to a more structured, intuitive package organization and clear separation between core and data access logic.
