+++ Blink3.Core/Repositories/Interfaces/IBlinkGuildRepository.cs
@@ -1,6 +1,6 @@
-using Blink3.DataAccess.Entities;
+using Blink3.Core.Entities;
 
-namespace Blink3.DataAccess.Interfaces;
+namespace Blink3.Core.Repositories.Interfaces;
 
 /// <summary>
 ///     Represents a repository for accessing and manipulating BlinkGuild entities.
