+++ Blink3.DataAccess/Repositories/UserTodoRepository.cs
@@ -1,5 +1,5 @@
-using Blink3.DataAccess.Entities;
-using Blink3.DataAccess.Interfaces;
+using Blink3.Core.Entities;
+using Blink3.Core.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
 namespace Blink3.DataAccess.Repositories;
