+++ Blink3.Web/Components/CreateTodoForm.razor
@@ -1,5 +1,5 @@
 @using Blink3.Web.Interfaces
-@using Blink3.DataAccess.Models
+@using Blink3.Core.DTOs
 <EditForm Model="_todo" OnValidSubmit="@(CreateTodoAsync)">
     <Modal @ref="_modal" OnHiding="OnModalHiding" Title="Create Todo">
         <BodyTemplate>
