+++ Blink3.Web/Components/TodoList.razor
@@ -1,5 +1,5 @@
-@using Blink3.DataAccess.Entities
-@using Blink3.DataAccess.Models
+@using Blink3.Core.Entities
+@using Blink3.Core.DTOs
 <ConfirmDialog @ref="dialog"/>
 
 @foreach (UserTodo todo in Todos)
