+++ Blink3.Web/Pages/Todo.razor
@@ -1,7 +1,7 @@
 @page "/Todo"
-@using Blink3.DataAccess.Entities
-@using Blink3.DataAccess.Models
+@using Blink3.Core.Entities
 @using Blink3.Web.Interfaces
+@using Blink3.Core.DTOs
 
 <PageTitle>Todo list</PageTitle>
 <h1 class="border-bottom mb-2 pb-2">Todo list</h1>
